Create Source Code
Fish keeping is a popular worldwide. Almost people from all 
the age groups like to keep aquarium in their home, offices etc 
for decoration purpose or as a hobby. 
• Fish keeping is not an easy job. We always need an aquarium or 
a pond for that. The hobby of fish keeping is broadly divided 
into three, freshwater, brackish, marine. 
• Among all these three, freshwater is considered to be the most 
popular hobby of keeping fish because it is easy to handle with 
freshwater fish and aquariums. It has always been difficult to 
take care of the fish and aquariums.
• We must feed the fish on time, we must maintain the 
temperature and also control the light, heater, and filter of the 
aquarium. All these steps are done manually. 
• The project, Smart aquarium has been designed by keeping in 
mind, the problem of those who cannot take care of their 
aquarium every day. 
• The feeder is powered by servo meter which on trigger of a 
button feeds the fish, sensor keeps an update of the temperature 
of the aquarium, and we can know the water level which helps 
in the growth of fish.